Refactor discovery module, add graph tests, narrow __init__ excepts (P1)
Completes the three P1 items from docs/code-quality-analysis.md:

1. Split action_evaluation/discovery.py (3001 LOC, 42 methods) into a
   discovery/ package with one mixin module per action family
   (line reconnection/disconnection, node splitting/merging, PST, load
   shedding, renewable curtailment, orchestrator) composed on top of a
   DiscovererBase that owns __init__, cached state, and shared helpers.
   Method bodies are copied verbatim via AST extraction — no logic change.
   Public import path (ActionDiscoverer) is preserved.

2. Add tests/test_graph_analysis.py covering inhibit_swapped_flows,
   get_n_connected_components_graph_with_overloads,
   identify_overload_lines_to_keep_overflow_graph_connected,
   get_subs_islanded_by_overload_disconnections, and
   get_graph_file_name, and tests/test_environment_pypowsybl.py covering
   file discovery, thermal-limit rescaling, and the compat wrappers.
   Both files rely on MagicMock/SimpleNamespace — no real env required.

3. Replace the bare except nest in __init__.py with narrowed exceptions
   (ImportError for the optional grid2op import, AttributeError for the
   monkey-patch) routed through logging instead of print-to-stderr.

Updates docs/code-quality-analysis.md with the P1 cleanup log.
