+++ expert_op4grid_recommender/__init__.py
@@ -13,27 +13,42 @@ graphs, applies expert rules to filter potential actions, and identifies
 corrective measures to alleviate line overloads.
 """
 
+import logging
+
 __version__ = "0.1.9"
+
+_logger = logging.getLogger(__name__)
 
 # --- GLOBAL MONKEY PATCH for grid2op 1.12+ compatibility ---
 # PyPowSyBlBackend 0.3.0 and other backends might not initialize self._sh_vnkv,
-# which is required by grid2op 1.12+ for grids with shunts.
+# which is required by grid2op 1.12+ for grids with shunts. The patch below
+# lazily initializes ``_sh_vnkv`` on the grid2op ``Backend`` base class the
+# first time ``get_shunt_setpoint`` is called.
+#
+# grid2op is optional: when it is not installed the patch is a no-op. Any other
+# failure is logged (debug level) to avoid surfacing noise at package import
+# time, but is not swallowed by a bare ``except``.
 try:
-    import sys
     import numpy as np
     import grid2op.Backend.backend as g2op_bk
-    _orig_get_shunt_setpoint = g2op_bk.Backend.get_shunt_setpoint
-    def _patched_get_shunt_setpoint(self):
-        if (not hasattr(self, "_sh_vnkv") or self._sh_vnkv is None) and getattr(self, "n_shunt", 0) > 0:
-            print(f"DEBUG: Initializing _sh_vnkv for {type(self)}", file=sys.stderr)
-            self._sh_vnkv = np.ones(self.n_shunt, dtype=np.float32) * 225.0
-        return _orig_get_shunt_setpoint(self)
-    g2op_bk.Backend.get_shunt_setpoint = _patched_get_shunt_setpoint
-    print("DEBUG: grid2op.Backend.Backend.get_shunt_setpoint patched", file=sys.stderr)
-except (ImportError, Exception) as e:
+except ImportError as exc:
+    _logger.debug("grid2op not available, skipping shunt-setpoint patch: %s", exc)
+else:
     try:
-        import sys
-        print(f"DEBUG: Failed to patch grid2op: {e}", file=sys.stderr)
-    except:
-        pass
+        _orig_get_shunt_setpoint = g2op_bk.Backend.get_shunt_setpoint
+
+        def _patched_get_shunt_setpoint(self):
+            if (
+                (not hasattr(self, "_sh_vnkv") or self._sh_vnkv is None)
+                and getattr(self, "n_shunt", 0) > 0
+            ):
+                _logger.debug("Initializing _sh_vnkv for %s", type(self))
+                self._sh_vnkv = np.ones(self.n_shunt, dtype=np.float32) * 225.0
+            return _orig_get_shunt_setpoint(self)
+
+        g2op_bk.Backend.get_shunt_setpoint = _patched_get_shunt_setpoint
+        _logger.debug("grid2op.Backend.Backend.get_shunt_setpoint patched")
+    except AttributeError as exc:
+        # grid2op present but unexpected API surface — log and move on.
+        _logger.debug("Failed to patch grid2op get_shunt_setpoint: %s", exc)
 # -----------------------------------------------------------
